+++ Changelog.md
@@ -21,6 +21,7 @@ Bugfixes:
  * References Resolver: Fix internal bug when using constructor for library.
  * Yul Optimizer: Make function inlining order more resilient to whether or not unrelated source files are present.
  * Type Checker: Disallow signed literals as exponent in exponentiation operator.
+ * Allow `type(Contract).name` for abstract contracts and interfaces.
 
 
 ### 0.7.0 (2020-07-28)